Revert "fix ci"
This reverts commit fa0cd9e2101242beccf8dfdd3c7cf9de506ce782.

+++ common/message/message_test.go
@@ -16,7 +16,7 @@ func TestAuthMessageSignAndVerify(t *testing.T) {
 	authMsg := &AuthMsg{
 		Identity: &Identity{
 			Name:      "testRoller",
-			Timestamp: uint64(time.Now().UnixNano()),
+			Timestamp: time.Now().UnixNano(),
 		},
 	}
 	assert.NoError(t, authMsg.Sign(privkey))
